backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
continuous-threading==2.0.6
fastapi==0.135.2
h11==0.16.0
idna==3.11
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyserial==3.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
